Enable Test_TC_OO_2_4 to run in CI (#25159)
We have additional logging in yamltest_with_chip_repl_tester.py that
will help us better understand the flake. Unfortunately we are not
able to reproduce the flake locally and need flake to happen in CI.

Additionally in the previous flake logs we do have the debug line
was not help as it was printing the address of the onOff argument
and not the value. That has been address in another PR that has since
been merged.