Remove unused "<access modifier>" bits from XML. (#26330)
Removing these does not change any codegen, either compile-time or for committed
files, so we're not using these for anything.  And they're not consistently
applied to the attributes that are fabric scoped, so if we did try to start
using them we would need to audit to decide where to do that, and not using them
anywhere makes that clear.

Fixes https://github.com/project-chip/connectedhomeip/issues/26316